Refactoring the project Structure, adding Mongo Client
