+++ src/main/java/com/udacity/course3/reviews/repositories/MySQL/ReviewRepository/ReviewRepository.java
@@ -1,4 +1,4 @@
-package com.udacity.course3.reviews.ReviewRepository;
+package com.udacity.course3.reviews.repositories.MySQL.ReviewRepository;
 
 import com.udacity.course3.reviews.entity.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
